Experimental new impl of CubicEOS in `CubicEOS2.hpp/cpp` now moved to `CubicEOS.hpp/cpp`
